# docs/swagger.yaml
components:
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: X-API-Key
      type: apiKey
info:
  title: API Gateway (KrakenD)
  version: 1.0.0
openapi: 3.0.0
paths:
  /orchestrateur/commande:
    post:
      summary: Créer une commande complète (saga orchestrée)
      description: |
        Lance une commande pour un client donné, en suivant une saga orchestrée :
        1. Création du client
        2. Création du panier
        3. Ajout du produit au panier
        4. Validation (checkout)
        En cas d’échec à une étape, un rollback est effectué automatiquement.
      parameters:
        - in: query
          name: produit_id
          required: true
          schema:
            type: integer
        - in: query
          name: magasin_id
          required: true
          schema:
            type: integer
        - in: query
          name: client_id
          required: true
          schema:
            type: integer
        - in: query
          name: quantite
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Succès – Commande traitée avec succès.
        '500':
          description: Échec – Une erreur est survenue durant la commande (rollback effectué).
      security:
        - ApiKeyAuth: []
      tags:
        - Orchestrateur
        
  /clients/:
    post:
      summary: Créer un client
      description: Crée un nouveau client dans le système.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                nom:
                  type: string
                prenom:
                  type: string
                email:
                  type: string
                adresse:
                  type: string
      responses:
        '200':
          description: Succès
      security:
      - ApiKeyAuth: []
      tags:
      - Clients
  /clients/{client_id}:
    get:
      summary: Récupérer un client par ID
      description: Récupère les informations d'un client spécifique.
      parameters:
      - in: path
        name: client_id
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Succès
      security:
      - ApiKeyAuth: []
      tags:
      - Clients
  /panier/:
    post:
      summary: Créer un panier
      description: Crée un nouveau panier pour un client. 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                client_id:
                  type: integer
      responses:
        '200':
          description: Succès
      security:
      - ApiKeyAuth: []
      tags:
      - Panier
  /panier/panier/{panier_id}:
    get:
      summary: Obtenir un panier par ID
      description: Récupère les détails d'un panier spécifique (tous les produits).
      parameters:
      - in: path
        name: panier_id
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Succès
      security:
      - ApiKeyAuth: []
      tags:
      - Panier
  /panier/panier/{panier_id}/checkout:
    post:
      summary: Valider le panier (checkout)
      description: Valide le panier et enregistre la commande.
      parameters:
      - in: path  
        name: panier_id
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Succès
      security:
      - ApiKeyAuth: []
      tags:
      - Panier
  /panier/{panier_id}/produit:
    post:
      summary: Ajouter un produit au panier
      description: Ajoute un produit spécifique au panier.
      parameters:
      - in: path  
        name: panier_id
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                magasin_id:
                  type: integer
                quantite:
                  type: integer
      responses:
        '200':
          description: Succès
      security:
      - ApiKeyAuth: []
      tags:
      - Panier
  /reporting/dashboard:
    get:
      summary: Afficher le tableau de bord
      description: Affiche les statistiques du tableau de bord (chiffre d'affaires, tendances, etc.).
      responses:
        '200':
          description: Succès
      security:
      - ApiKeyAuth: []
      tags:
      - Rapports
  /reporting/ruptures:
    get:
      summary: Afficher les ruptures de stock
      description: Affiche les produits en rupture de stock.
      responses:
        '200':
          description: Succès
      security:
      - ApiKeyAuth: []
      tags:
      - Rapports
  /reporting/ventes:
    get:
      summary: Afficher les ventes
      description: Affiche les ventes effectuées.
      responses:
        '200':
          description: Succès
      security:
      - ApiKeyAuth: []
      tags:
      - Rapports
  /restock/stock_par_magasin:
    get:
      summary: Afficher le stock par magasin
      description: Affiche le stock disponible pour un magasin spécifique.
      parameters:
      - in: query
        name: magasin_id
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Succès
      security:
      - ApiKeyAuth: []
      tags:
      - Réapprovisionnement
  /restock/transferer:
    post:
      summary: Transférer du stock entre magasins
      description: Transfère une quantité spécifique d'un produit d'un magasin source vers un magasin destination.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                produit_id:
                  type: string
                quantite:
                  type: integer
                magasin_source_id:
                  type: integer
                magasin_destination_id:
                  type: integer
      responses:
        '200':
          description: Succès
      security:
      - ApiKeyAuth: []
      tags:
      - Réapprovisionnement
  /stock/:
    get:
      summary: Récupérer tous les produits en stock
      description: Récupère la liste de tous les produits disponibles en stock.
      responses:
        '200':
          description: Succès
      security:
      - ApiKeyAuth: []
      tags:
      - Stock
  /stock/annulation:
    post:
      summary: Annuler une vente de produit
      description: Annule une vente de produit en spécifiant l'ID du produit, la quantité et l'ID du magasin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                produit_id:
                  type: string
                quantite:
                  type: integer
                magasin_id:
                  type: integer
      responses:
        '200':
          description: Succès
      security:
      - ApiKeyAuth: []
      tags:
      - Stock
  /stock/rechercher:
    get:
      summary: Rechercher un produit
      description: Recherche un produit par terme dans le nom ou la catégorie ou le id.
      parameters:
      - in: query
        name: terme
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Succès
      
      security:
      - ApiKeyAuth: []
      tags:
      - Stock
  /stock/update/{produit_id}/magasin/{magasin_id}:
    put:
      summary: Mettre à jour le stock d'un produit dans un magasin
      description: Met à jour les informations d'un produit spécifique dans un magasin.
      parameters:
      - in: path
        name: produit_id
        required: true
        schema:
          type: integer
      - in: path
        name: magasin_id
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nom:
                  type: string
                categorie:
                  type: string
                prix:
                  type: float
                quantite:
                  type: integer
      responses:
        '200':
          description: Succès
      security:
      - ApiKeyAuth: []
      tags:
      - Stock
  /stock/vente:
    post:
      summary: Enregistrer une vente de produit
      description: Enregistre une vente de produit en spécifiant l'ID du produit, la
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                produit_id:
                  type: string
                quantite:
                  type: integer
                magasin_id:
                  type: integer
      responses:
        '200':
          description: Succès
      security:
      - ApiKeyAuth: []
      tags:
      - Stock
security:
- ApiKeyAuth: []
servers:
- url: http://localhost:8082